+++ src/days/day04.rs
@@ -1,6 +1,6 @@
 use crate::utils::VecStringTrim;
 
-pub fn s1(input: Vec<String>) -> i32 {
+pub fn s1(input: Vec<String>) -> i64 {
     let input = input.foreach_trim();
     let mut accum = 0;
 
@@ -50,7 +50,7 @@ pub fn s1(input: Vec<String>) -> i32 {
     accum
 }
 
-pub fn s2(input: Vec<String>) -> i32 {
+pub fn s2(input: Vec<String>) -> i64 {
     let input = input.foreach_trim();
 
     let mut accum = 0;
